Fix bug in code sending new peep to server
The way the code was set up was causing a race condition between
the sending of the new peep and retrieving the list of peeps. This
meant that sometimes the list was retrieved before the new peep
was added to the database.

This was fixed by moving the callback function call into the then
part of the axios call. This ensures that it only gets called when
the client has received the response from the server after adding
the new peep.