Add ProtectedRoute component and useUserState hook
The useUserState hook provides a convenience method to check if a
user is logged in.

ProtectedRoute is used to prevent the user accessing certain
routes unless they are logged in. At the moment the only route
that has been protected in this way is /peeps. In all cases, you
are redirected to the login route if you try to access a protected
route without logging in.