Refactor contexts to simplify where frontend GETS from backend
This was a significant refactor that removed the need for the
Peep component to get individual peep details from the backend.
Instead, it accesses the MainContext, which holds the complete
peeps list.

This change required significant refactors across the board, mainly
in the tests, a lot of which had to be wrapped in contexts and
have axios mocks applied.